backend/requirements.txt
bs4
openai
langchain
langchain-community
langchain-chroma
langchain-openai
tqdm
flask
flask-cors
google-auth
google-cloud-secret-manager